feat(trainingAndQualsHowToTab): Move generateHowToTab into separate file
